Switched package to prepublish
The infra publishing lifecycle requires that we build a tarball
without the invocation of the actual 'publish' command. That
tarball is then uploaded to static.openstack.org. Switching from
the custom 'package' script to the 'prepublish' script hooks into
the 'npm pack' lifecycle, which will allow us to minify/package
before sending it to static.

Change-Id: If695c5a29b90e0019f7651a542fced602ace0383