Removed X-Client header
This header is mostly out of courtesy to the API, however to make
the client work 'out of the box' it requires additional configuration
on the API side, which is not well documented, or desired. Instead,
I am opting to remove it, and allow third-party extension of the app
to add what headers they themselves require.

Change-Id: I40b06d50c7aaeab01c1c908557228920126d0ea8