Split storage components into individual files
This patch asserts the "One file per thing" best practice
for the storage services. It also removes the $$secureStorage
component, since it doesn't actually exist, or is a thing that
browsers actually do.

The behavior of the storage API has also been changed. Instead
of returning 'null' when a key is not found, it expects a return of
'undefined'. Unit tests to assert all edge cases have been added,
and the coverage threshold for the webclient has been raised.

Change-Id: I7ddfbc92292d4bded0a3f1208aeecb1ab89eeac3