Expanded clean command.
This command extends the gulp 'clean' command to scrubbing the
entire project. It's intended for developer use only, in cases where
switching branches requires different dependencies and/or project
structure.

Note that this command will require a subsequent invocation of
'npm install' to return the project to a usable state.

Change-Id: I3b884b98c91617e252b8b373e04e2f019422c1b8