Added $$resourceCache service.
This patch adds a simple class that may act as a resource cache,
functionally separate but mostly identical to the $$memoryStorage
mechanism. Noted differences is that $$resourceCache does not
serialize its content, and by using it we do not pollute each others'
name spaces.

This deprecates the $$resourceFactory, as it turned out to be a
little overengineered. A much simpler caching mechanism will be
far more effective. See subsequent patches for examples.

 Tests are included, and coverage threshold has been updated.

Change-Id: I5f19e499062a1395f8c8109e48bbcf5724b0e233