Configuration Resource
This splits configuration management into two components, each with
a separate purpose. The first, $$configuration, acts as an ngResource-style
API abstraction that loads and maintains the list of currently active
configurations. The second, $$selectedConfiguration, provides the
mechanism by which a specific configuration is selected and validated.

New features added:
- A configuration may be force-loaded into the $$configuration provider
  during app initialization.
- The autodetection mechanism has been renamed to 'default'.
- The localStorage, configFile, and default detection mechanism may
  now be enabled or disabled (default disabled).

The main drive for this is separation of concerns and testability. When
writing tests for downstream API's, having to navigate the complex chain of
promise resolutions simply to load an API test configuration was
obnoxiously difficult. Separating selection and loading greatly simplifies
this. Furthermore, adopting a more ngResource-style resolution of
resources - where the object returned from a CRUD request will be the actual
data object in use, also simplifies usage in the user interface.

Tests have been provided, and testing threshold has been updated accordingly.

Eslint was updated as well, as the indentation rules did not seem to work
too well with functions in function parameters being declared in a second
line.

Change-Id: I345b6f8e5da9d3064eabb3cc8cf0113e59626ac5