Separated UI and Configuration Logic
This removes the use of UI components from the openstack library,
making it reusable outside of ironic-webclient. It is mainly achieved by
removing the addLocal invocation of $modal from $$configuration, nad
moving the ConfigurationAddController from the openstack module
to the ironic root application module. Tests for the
ConfigurationAddController have been also added, and the test
threshold has been raised accordingly.

Change-Id: Icfd7504be966cee330a934c99b50465fd8ff7993