Introduced Root State
The introduction of a root state permits us to move the overall
application layout to a location that is common to the entire
application. It reduces template duplication, and allows us
to specifically target view containers at any level of the
ui framework.

The ApplicationController has been renamed to describe its
new function, and a new 'main.html' template was added.
Note that there are quite a few whitespace changes, as
the indentation of many files changed when the encapsulating
div container moved to the global template.

Change-Id: I2d0b7147903f0af0560cd33903a1e6e4df309f55