Enroll Node
This patch creates a very simple "enroll node" process, which
permits the selection of a driver, as well as setting that drivers'
properties based on the results received from the API. It adds
an IronicDriverProperties resource to expose individual driver
properties to the UI. Unfortunately the driver properties do not
contain type, validation, or required data in a machine-readable
fashion, so these are not applied to the form generation.

Editing a driver's properties, or un-enrolling a node, is not yet
supported.

Change-Id: Idd7015937f908b5567717a8d34fa4958ac939041