Added Ironic API result transformation.
When the ironic API returns an error, it is actually a JSON object
with a single field named "error_message", which contains a JSON
encoded object (doubly encoded) with the actual error message.
This must be decoded twice to be easily accessible to the UI.

When the ironic API returns a list response, it is actually a JSON
object with a single field named as the resource that was requested.
This needs to be reduced, so that the list response is an array rather
than an object.

This patch adds a result transformation chain, independent of
ironic resources, that applies all likely transformations in a chain.
Tests have been provided.

Change-Id: I4deed08fe3f3a65f4f48b54620c791f0c3605238