Added API error handling.
When the ironic API returns an error, it is actually a JSON object
with a single field named "error_message", which contains a JSON
encoded object (doubly encoded) with the actual error message.

This patch ensures that the internal object is properly decoded
before it is passed through to the UI. Tests have been provided.

Change-Id: I4deed08fe3f3a65f4f48b54620c791f0c3605238