Enroll Node
This patch creates a very simple "enroll node" process, which
permits the selection of a driver, as well as setting that drivers'
properties based on the results received from the API.

New generic mocks to simulate an API have been added to support
necessary tests, and coverage thresholds have been updated.

Editing a driver's properties, or un-enrolling a node, is not yet
supported.

Change-Id: Idd7015937f908b5567717a8d34fa4958ac939041