Collapsed node module
The nodes module had no real reason to exist separately from the
existing root module. I've removed it, and transferred the relevant
states and controllers to the root application module.

Also, this patch cleans up the index.html file, which tried to load
to javascript files that it shouldn't have.

Change-Id: I034aadaa2339b348fb67a170c789e11cf77fe104