Basic node action hooks.
This patch adds UI rendering in the node_list of actions available
to each node based on its state. It then adds a simple switching
mechanism that will trigger different modals based on the desired
action, and a placeholder controller which informs the user that
development is ongoing.

Eslint rules had to be updated, because ng_no_services prevents
usage of $q in anything but services.

Change-Id: I77b441d84a1b92d359879ea99b38804decaae9e9