Adjusted flow for header elements
The header components now flow in accordance to the approved
UI. The search bar extneds to the entire remaining width of the
screen, and small adjustments have been made to the title
to offer more space in smaller form factors. It was necessary
to move the branding into header.html, in order to permit
proper column sizing (nesting them created weird padding issues).

Note that this UI doesn't work too well on a mobile screen -
that experience should probably receive its own treatment.

http://git.openstack.org/cgit/openstack/openstack-ux/plain/mocks/Ironic%20Standalone%20-%2007%20March%202016.pdf

Change-Id: I1b1f8f9bb4f9f9e79b86260be3a554274d8c44ed