Available node transitions now available in node_list
This patch ensures that available provisionState transitions,
as well as available powerState transitions, are accessible to the
user in the actions dropdown. The default action has been set
to whichever power value is first in the list. The control buttons -
dropdown and reboot - are only enabled if at least one node is selected,
though no filtering is performed to detect which actions are available.

Change-Id: I8cc91411a5d0ddcee6e1bfeddbca069594ff6dae