Basic node action hooks.
This patch adds action handling for power and provision state
changes in the node list view, via the NodeActionController. Or,
in simpler terms, if you select a node and an action, a modal will
now pop up to tell you that particular action is not yet supported.
Actions will be added in subsequent patches.

The NodeActionController itself has been greatly simplified,
because it needs to serve as both a single-node and a multi-node
action handler; in other words, the previous opinionated
implementation of one-handler-per-node has been genericized.

Change-Id: I77b441d84a1b92d359879ea99b38804decaae9e9